fix: compare costs as numbers not strings

+++ src/query/Query.js
@@ -218,7 +218,9 @@ export default class Query extends Executable {
             if (this._queryPayment == null) {
                 const actualCost = await this.getCost(client);
 
-                if (cost.toTinybars() < actualCost.toTinybars()) {
+                if (
+                    cost.toTinybars().toInt() < actualCost.toTinybars().toInt()
+                ) {
                     throw new MaxQueryPaymentExceeded(cost, actualCost);
                 }
